### Description
Brief: I have production case that uses big amount of simple policies (~100K) that need to have enforcement time < 50ms. 
This use case have low performance at casbin-cpp. Casbin-cpp cashing not comes the solution - on uniform distribution of request. Cash filling is quite long. 

I have implemented unordered_set storage for policies for "basic model with allow only" to be used on enforcement.

--==Solution have **performance gain X185 times**. ==--

This gain will not depend on policies count anymore - O(1) complexity of traverse on enforcement:

Before the modification:
```
RUNNING BENCHMARK
Benchmark                                                     Time                   CPU                           Iterations
BenchmarkBasicModelLargeSize                29929285 ns     29822588 ns           85
```
After the modification:
```
Benchmark                                                     Time                   CPU                           Iterations
BenchmarkBasicModelLargeSize                 161366 ns         161240 ns                  4454
```


Workaround:
Traversing of policies on enforcement at casbin-cpp implemented for vector. That quite long on using not first (but last ) element of the vector. It nice to reflect this on benchmark.

Solution:
1) Instead of vectors use unordered_set to store the policies. It makes searching for specific policy complexity to be O(1) instead of O(n). 
2) Improve only "basic model with allow only" case
3) Filter of policies ( stored in unordered_set ) if "basic model with allow only" used to get just one element for enforcement
4) Put filtered collection to main enforce algorithm
5) Use option of cmake to have storage of policies as vector or unordered_set ( to reduce memory footprint on specific use cases )

Provements:
1) please come at build/ directory and run build_actions.sh
```bash
#!/usr/bin/env bash
echo INITIAL CLEANING
git checkout master
cd build
cmake ../
cmake --build . --target clean > /dev/null 2>&1
echo COMPILATION
cmake --build . -j 8 > /dev/null 2>&1
echo RUNNING BENCHMARK
./tests/benchmarks/casbin_benchmark 2> /dev/null | grep Benchmark

echo SECONDARY CLEANING
git checkout AlexeyChashchegorov:master
cmake ../
cmake --build . --target clean > /dev/null 2>&1
echo COMPILATION
cmake --build . -j 8 > /dev/null 2>&1
echo RUNNING BENCHMARK
./tests/benchmarks/casbin_benchmark 2> /dev/null | grep Benchmark
```

Results of running:
```
INITIAL CLEANING
SET BUILD ENV WITH BUILD_HASHED=OFF
COMPILATION
RUNNING BENCHMARK
Benchmark                                          Time             CPU   Iterations
BenchmarkRaw                                     165 ns          165 ns      4248888
BenchmarkBasicModel                           134863 ns       134836 ns         5390
BenchmarkBasicModelLargeSize                29929285 ns     29822588 ns           85
BenchmarkRBACModel                            205365 ns       204697 ns         4116
BenchmarkRBACModelSizesSmall                 2527201 ns      2526538 ns          649
BenchmarkRBACModelSmall                      4573460 ns      4571217 ns          276
BenchmarkRBACModelWithResourceRoles           148069 ns       148069 ns         4886
BenchmarkRBACModelWithDomains                 185278 ns       185208 ns         3848
BenchmarkKeyMatchModel                        162665 ns       162610 ns         4428
BenchmarkRBACModelWithDeny                    288470 ns       288407 ns         3060
BenchmarkPriorityModel                        145488 ns       145451 ns         4961
BenchmarkCachedRaw                               165 ns          165 ns      4232088
BenchmarkCachedBasicModel                        737 ns          721 ns      1004650
BenchmarkCachedRBACModel                         710 ns          703 ns       980626
BenchmarkCachedRBACModelSmall                    616 ns          615 ns      1101599
BenchmarkCachedRBACModelWithResourceRoles        692 ns          692 ns       995138
BenchmarkCachedRBACModelWithDomains              685 ns          679 ns      1021048
BenchmarkCachedKeyMatchModel                     628 ns          628 ns      1105758
BenchmarkCachedRBACModelWithDeny                 690 ns          690 ns       997137
BenchmarkCachedPriorityModel                     685 ns          685 ns      1013729
BenchmarkVectorOperations                        493 ns          493 ns      1425676
BenchmarkHasPolicySmall                        88897 ns        88870 ns         8001
BenchmarkAddPolicySmall                       452076 ns       451889 ns         2019
BenchmarkRemovePolicySmall                      1580 ns         1578 ns       442165
BenchmarkRoleManagerSmall                      96288 ns        96244 ns         7281
SECONDARY CLEANING
SET BUILD ENV WITH BUILD_HASHED=ON
COMPILATION
RUNNING BENCHMARK
Benchmark                                          Time             CPU   Iterations
BenchmarkRaw                                     200 ns          200 ns      3460704
BenchmarkBasicModel                           161802 ns       161720 ns         4865
BenchmarkBasicModelLargeSize                  161366 ns       161240 ns         4454
BenchmarkRBACModel                            173449 ns       173367 ns         4506
BenchmarkRBACModelSizesSmall                 2767169 ns      2765977 ns          650
BenchmarkRBACModelSmall                      4565711 ns      4564597 ns          278
BenchmarkRBACModelWithResourceRoles           212746 ns       212691 ns         3864
BenchmarkRBACModelWithDomains                 282814 ns       282422 ns         2806
BenchmarkKeyMatchModel                        304304 ns       304051 ns         2785
BenchmarkRBACModelWithDeny                    288726 ns       288308 ns         3057
BenchmarkPriorityModel                        291688 ns       291589 ns         3004
BenchmarkCachedRaw                               196 ns          196 ns      3597289
BenchmarkCachedBasicModel                        687 ns          687 ns       994912
BenchmarkCachedRBACModel                         691 ns          690 ns       996895
BenchmarkCachedRBACModelSmall                    636 ns          636 ns      1065336
BenchmarkCachedRBACModelWithResourceRoles        689 ns          688 ns      1007484
BenchmarkCachedRBACModelWithDomains              690 ns          690 ns      1004895
BenchmarkCachedKeyMatchModel                     636 ns          636 ns      1084700
BenchmarkCachedRBACModelWithDeny                 685 ns          685 ns      1012922
BenchmarkCachedPriorityModel                     687 ns          686 ns      1012717
BenchmarkVectorOperations                        493 ns          493 ns      1417093
BenchmarkHasPolicySmall                        93203 ns        90066 ns         7917
BenchmarkAddPolicySmall                       415859 ns       415684 ns         2125
BenchmarkRemovePolicySmall                      1633 ns         1632 ns       429664
BenchmarkRoleManagerSmall                      96992 ns        96944 ns         7241
```

### Screenshots/Testimonials

Actions on the PR:
- using of unordered_set at assertion_map to prototype the fast usage
- implement compilation flag to select vector/unordered_set as storage type of policies at assertion_map
- implement basic_model detector: to know when unordered_set can be used
- to fix test failure on "priority model" - addition of abstract collection "PoliciesValues" that hides behavior of vector/unordered_set and unify the usage. "Priority model" only uses vector as policies storage and it can not been changed any possible way.
- construction of collection depends on model: basic model -> PoliciesValues will be based on unordered_set
- filtering of model policies at assertion map depends on PoliciesValues internal collection type: if "is_hash()" of the collection returns true - find at collection and hold just value that match.
- moving out of the compilation flag ( automated selection of collection makes it possible )