@hsluoyz 2024-03-06
@AlexeyChashchegorov fix CI:

![image](https://github.com/casbin/casbin-cpp/assets/3787410/f11b1d20-59ab-4ba2-861d-0cf9546d2f9f)