@AlexeyChashchegorov 2024-03-06
I will figure out why it fails. My local build on osx passed with no problems.