@AlexeyChashchegorov 2024-03-13
PR description updated.
Code has significant update:
- abstract collection PoliciesValues to make vector / unordered_set be threat uniform way. 
- all tests passed
- Priority model is only blocker to do not use unordered_set directly at each of the models. Priority model relies on order user defined - it must be hold as vector