@AlexeyChashchegorov 2024-03-14
several comments:
- it is possible to use unordered_set for grouping policies as well ( speedup will be the same and not dependent of policies count) 
- it is possible to use deny effects for basic and rbac models as well with minimal limitations
- I can and would like to make this changes as well but not in this PR.