@hsluoyz 2024-03-14
@AlexeyChashchegorov plz fix CI errors:

![image](https://github.com/casbin/casbin-cpp/assets/3787410/8c0e551d-decb-4f04-9bf8-04559e754214)