@AlexeyChashchegorov 2024-03-15
I have just 20.04.6 and g++9.4 not that used on CI. But it pretty same to fix most of the bugs on linux and have unittests passed. 

I have some problems on build of py bindings. But I believe it may be compiler specific problem. 
Let's look.