@hsluoyz 2024-03-15
@AlexeyChashchegorov plz fix:

![image](https://github.com/casbin/casbin-cpp/assets/3787410/75ca3318-859b-41dd-adbc-c4053fdde864)