@AlexeyChashchegorov 2024-03-16
I will look at CI. Some time needed to get Win host available to compile. 

@hsluoyz If you can direct me how to get proper Win host / pod with appropriate compiler to debug - I will be very appreciative.  ( On my side it might consume more time )