@AlexeyChashchegorov 2024-03-17
It should works for Windows as well now.